add: home page product lising
