+++ src/context/listingContext.js
@@ -1,5 +1,7 @@
-import { React, createContext, useContext, useReducer } from "react";
+import { React, createContext, useContext, useReducer, useEffect } from "react";
+import { maxPriceReducer } from "../helpers/maxPriceReducer";
 import { listingReducer } from "../reducers/listingReducer";
+import { getAllProducts } from "../services/products";
 
 const ListingContext = createContext();
 
@@ -9,7 +11,18 @@ export const ListingProvider = ({ children }) => {
     categoryList: [],
     price: 0,
     rating: 1,
+    maxPrice: 0,
   });
+
+  useEffect(async () => {
+    const res = await getAllProducts();
+    const data = await res.data;
+    const maxPrice = data.products.reduce(maxPriceReducer, 0);
+    listingDispatch({
+      type: "PRODUCT_LIST_SET",
+      payload: { products: data.products, price: Number(maxPrice) },
+    });
+  }, []);
   return (
     <ListingContext.Provider value={{ listingState, listingDispatch }}>
       {children}
